add comment to function

+++ andrew.sh
@@ -1,16 +1,26 @@
 #!/bin/sh
 
-cd ~/corewar/corewar
-make
+EOC='\033[0m' # End Of Color
+RED='\033[0;31m'
+YEL='\033[1;33m'
+GRE='\033[0;32m'
+BLU='\033[0;34m'
 
-rm result.txt result2.txt
+echo "${BLU}Running ${RED}Androunette Corewar : ${EOC}"
 
-cycle=5500
+cd ~/corewar/corewar #path from home to corewar exec
 
+if [[ -s result.txt || -s result2.txt ]] ; then
+	rm result.txt result2.txt
+fi ;
+
+cycle=5500 # Start Dumping...
+
+#put 1 till 4 champion(s) path from corewar
 champ1='../champion/Fuck_Eat_Repeat.cor'
 champ2='../ressources/champs/Gagnant.cor'
 champ3='../ressources/champs/Octobre_Rouge_V4.2.cor'
-champ4=rien
+champ4=nothing
 
 rm_color () {
 	sed 's/\^\[\[0m//g' $1 > result.txt
@@ -20,7 +30,7 @@ rm_color () {
 	sed 's/.$/ /g' $1 > result.txt
 }
 
-while test $cycle != 12500
+while test $cycle != 12500 # End Cycle...
 	do
 	./corewar -d $cycle -n -1 $champ1 \
 			   -n -2 $champ2 \
@@ -36,7 +46,9 @@ while test $cycle != 12500
 	rm result3.txt
 	diff result.txt result2.txt > check
 	if [[ -s check ]] ; then
-		echo Au cycle $cycle
+		echo "${RED}KO ${EOC}cycle ${RED}$cycle${EOC}"
 	fi ;
 	cycle=$(($cycle + 100))
 done
+
+rm result.txt result2.txt
